refactor: :art: Refatorado payment para novo padrao

+++ FIAP/FIAP.Diner.Domain/Checkout/IExternalPaymentService.cs
@@ -2,6 +2,6 @@ namespace FIAP.Diner.Domain.Checkout
 {
     public interface IExternalPaymentService
     {
-        Task<(string, string)> GenerateQRCode(double amount);
+        Task<(string, string)> GenerateQRCode(decimal amount);
     }
 }
